refactor(mobile): move use-live-swap-estimate to swap-state/hooks
