+++ apps/mobile/src/features/swap/components/quote-preview/quote-preview.tsx
@@ -1,6 +1,6 @@
 import Animated, { Easing, FadeIn, FadeOut } from 'react-native-reanimated';
 
-import { LiveSwapEstimate } from '@/features/swap/hooks/use-live-swap-estimate';
+import { LiveSwapEstimate } from '@/features/swap/swap-state/hooks/use-live-swap-estimate';
 import { SwapState } from '@/features/swap/swap-state/swap-state.types';
 
 import { assertExistence, assertUnreachable } from '@leather.io/utils';
