+++ apps/mobile/src/i18n/locales/en/messages.po
@@ -558,7 +558,7 @@ msgstr ""
 #: src/features/approver/components/nonce-sheet.tsx:18
 #: src/features/send/components/memo.tsx:88
 #: src/features/send/components/recipient/recipient-guard-sheet.tsx:30
-#: src/features/swap/screens/swap-review-screen.tsx:207
+#: src/features/swap/screens/swap-review-screen.tsx:210
 #: src/features/wallet-manager/recover-wallet/recover-wallet-sheet.tsx:22
 msgid "Confirm"
 msgstr "Confirm"
@@ -792,7 +792,7 @@ msgid "Enter your Secret Key"
 msgstr "Enter your Secret Key"
 
 #: src/features/swap/components/quote-preview/quote-preview-content.tsx:53
-#: src/features/swap/screens/swap-review-screen.tsx:191
+#: src/features/swap/screens/swap-review-screen.tsx:194
 msgid "Estimated fees"
 msgstr "Estimated fees"
 
@@ -1197,7 +1197,7 @@ msgstr "Locking"
 msgid "Mainnet, testnet or signet"
 msgstr "Mainnet, testnet or signet"
 
-#: src/features/swap/screens/swap-review-screen.tsx:202
+#: src/features/swap/screens/swap-review-screen.tsx:205
 msgid ""
 "Make sure everything looks correct.\n"
 "Confirmed transactions cannot be undone."
@@ -1254,7 +1254,7 @@ msgstr "Memo updated"
 msgid "Message"
 msgstr "Message"
 
-#: src/features/swap/screens/swap-review-screen.tsx:161
+#: src/features/swap/screens/swap-review-screen.tsx:164
 msgid "Min. receive"
 msgstr "Min. receive"
 
@@ -1431,7 +1431,7 @@ msgid "Price"
 msgstr "Price"
 
 #: src/features/swap/components/price-impact-info-sheet.tsx:8
-#: src/features/swap/screens/swap-review-screen.tsx:182
+#: src/features/swap/screens/swap-review-screen.tsx:185
 msgid "Price impact"
 msgstr "Price impact"
 
@@ -1485,7 +1485,7 @@ msgid "Rarity rank"
 msgstr "Rarity rank"
 
 #: src/features/swap/components/quote-preview/quote-preview-content.tsx:39
-#: src/features/swap/screens/swap-review-screen.tsx:146
+#: src/features/swap/screens/swap-review-screen.tsx:149
 msgid "Rate"
 msgstr "Rate"
 
@@ -1599,7 +1599,7 @@ msgstr "Reveal account"
 msgid "Review"
 msgstr "Review"
 
-#: src/features/swap/screens/swap-review-screen.tsx:71
+#: src/features/swap/screens/swap-review-screen.tsx:74
 msgid "Review Swap"
 msgstr "Review Swap"
 
@@ -1752,7 +1752,7 @@ msgstr "SIP-010"
 msgid "Skip for now"
 msgstr "Skip for now"
 
-#: src/features/swap/screens/swap-review-screen.tsx:169
+#: src/features/swap/screens/swap-review-screen.tsx:172
 msgid "Slippage"
 msgstr "Slippage"
